office Controller更新List，Info接口

+++ src/main/java/com/bo/AvailableOfficeBO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class AvailableOfficeBO {
-    Long officeId;
+    Integer officeId;
     String officeName;
 }
